adds tests for hocs
