+++ src/components/offers-main/offers-main.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import PlaceCardList from '../place-card-list/place-card-list.jsx';
 import {connect} from 'react-redux';
-import {getSortedOffers} from '../../reducers/data/selectors.js';
+import {getSortedOffers} from '../../reducers/data/reselectors.js';
 import {PlaceCardType} from '../../const/const.js';
 
 const OffersMain = ({offers, isAuth}) => {
