+++ src/components/page-main/page-main.jsx
@@ -10,7 +10,8 @@ import Header from '../header/header.jsx';
 import {connect} from 'react-redux';
 import {getAuthorizationStatus} from '../../reducers/user/selectors.js';
 import {getSortType, getActiveCity} from '../../reducers/app/selectors.js';
-import {getCities, getCityOffersCoords} from '../../reducers/data/selectors.js';
+import {getCities} from '../../reducers/data/selectors.js';
+import {getCityOffersCoords} from '../../reducers/data/reselectors.js';
 import withOpenState from '../../hocs/with-open-state/with-open-state.jsx';
 import withLoading from '../../hocs/with-loading/with-loading.jsx';
 import withPageError from '../../hocs/with-page-error/with-page-error.jsx';
