+++ src/reducers/data/selectors.js
@@ -1,6 +1,3 @@
-import {createSelector} from 'reselect';
-import {sortOffers} from '../../util.js';
-import {getActiveCityName, getSortType} from '../app/selectors.js';
 import NameSpace from '../name-space.js';
 
 export const getOffers = (state) => state[NameSpace.DATA].offers || [];
@@ -10,38 +7,3 @@ export const getComments = (state) => state[NameSpace.DATA].comments || [];
 export const getFavorites = (state) => state[NameSpace.DATA].offers.filter((it) => it.isFavorite);
 
 export const getNearbyCoordList = (state) => getNearbyList(state).map((it) => it.coord);
-
-export const getSortedOffers = createSelector(
-    [
-      getOffers,
-      getActiveCityName,
-      getSortType
-    ],
-
-    (offers, activeCityName, sortType) => {
-      if (offers.length === 0) {
-        return [];
-      }
-
-      const selectedOffers = activeCityName !== null ? offers.filter((it) => it.city === activeCityName) : offers;
-
-      return sortOffers(selectedOffers, sortType);
-    }
-);
-
-export const getCityOffersCoords = createSelector(
-    [
-      getOffers,
-      getActiveCityName,
-    ],
-
-    (offers, activeCityName) => {
-      if (offers.length === 0) {
-        return [];
-      }
-
-      const selectedOffers = activeCityName !== null ? offers.filter((it) => it.city === activeCityName) : offers;
-
-      return selectedOffers.map((it) => it.coord);
-    }
-);
